changed: dvdplayer now calculate timestamps using double datatype instead of int64. changed: clenaup of stale code in dvdplayer and some reorganization
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk/XBMC@10427 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/cores/dvdplayer/DVDOverlayContainer.cpp
@@ -77,7 +77,7 @@ CDVDOverlay* CDVDOverlayContainer::Remove(CDVDOverlay* pOverlay)
   return pNext;
 }
 
-void CDVDOverlayContainer::CleanUp(__int64 pts)
+void CDVDOverlayContainer::CleanUp(double pts)
 {
   CDVDOverlay* pOverlay = NULL;
